requirements.txt
# Package requirements

Flask ~= 1.0
mysql-connector ~= 2.2
